Wrote some _absolutely_ cursed filtering
The process of exporting and sending the Docker images is time consuming
and, in the spirit of Ansible's idempotency, should only be done when
needed. Hence, I slapped together some "delightful" filtering code to
check if our Docker images already exist on the host and, if so,
skipping to transfer them.

The language of Ansible filters is not a nice one. I've done my best
here to make it reasonably readable, but... yeah. It is what it is.